requirements.txt
asgiref==3.5.0
cfgv==3.3.1
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
filelock==3.4.2
identify==2.4.8
nodeenv==1.6.0
platformdirs==2.4.1
pre-commit==2.17.0
psycopg2==2.9.3
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.13.1
